BTDScripts/requirements.txt
GitPython==3.1.43
colorama==0.4.6